Moved Unicode namespace out of GeneralTextHandler file
