+++ client/renderSDL/CBitmapFont.cpp
@@ -14,7 +14,7 @@
 
 #include "../../lib/vcmi_endian.h"
 #include "../../lib/filesystem/Filesystem.h"
-#include "../../lib/CGeneralTextHandler.h"
+#include "../../lib/TextOperations.h"
 #include "../../lib/Rect.h"
 
 #include <SDL_surface.h>
